[INTERNAL] StateUtil: fixes for diff calculation
- bug fix in SelectionController: before only for move changes the index
has been respected, but in case an item will be added in the same turn
as it will be repositioned, there is only an add change. This has been
adjusteed to take this case into account.

- bug fix for xConfigAPI: fixes that only the last applied column width is
taken into account, since the aggregations in the object overwrote the
earlier value

BCP: 2380005290
BCP: 2380004788
BCP: 2380004855
Change-Id: I53bfa4afc37abf7893a8848f74554bd10039826a