[FIX] sap.m.SinglePlanningCalendarMonthGrid: More link shows correct number of appointments
Issue:
Appointments spanning multiple days, that are not rendered on the start
day, cause errors in calculating 'x more' appointments for the rest of
the days.

Solution:
The more link is now calculated based on how many appointments are
hidden below the maximum number of appointments that can be rendered on
that day.

BCP: 2270156096
Change-Id: If201dd50c4454c9f83f5fd7ee0414fbac504a61c