[FEATURE] SimpleForm: use ResponsiveGridLayout as default
as ResponsiveLayout is deprecated for a long time and cannot be used in Legacy-free UI5.

ColumnLayout is the current layout defined by UX. For most simple use-cases is should look the same like before.
But as it might break unsupported usages (nested layoutes...) now the default is changes to ResponsiveGridLayout.
In UI5 version 2.0 the ColumnLayout should be used as default.

Change-Id: Ie0c6dd734830f05aaea0ab679cb9dd968e0148c2
JIRA: CPOUIFDENVER-1867