[INTERNAL] ui5loader: break up module execution into tasks
The async implementation of the loader APIs is partially based on
promises. For larger trees of preloaded modules, this can result in
equally large sequences of micro-task executions for the module factory
functions. Such micro-task sequences contribute to the execution time of
the enclosing browser task and thereby block the main thread.

To reduce this blocking behavior, the loader now has the possibility to
trigger new tasks when the elapsed time of the current task exceeds a
configurable threshold.

For now, the new scheduling mechanism is off by default. If further
measurements show a significant improvement and no regressions, the
scheduling might become default behavior in future.

Until then, the scheduling can be activated with the experimental URL
parameter

  ?sap-ui-xx-maxLoaderTaskDuration=<n>

where <n> is a number between -1 and Infinity, representing a task
duration (in milliseconds) after which new module executions will be
postponed.

Note that the scheduling only affects the execution of preloaded module
factories. Other micro-task sequences are not affected and still might
cause blocking.

Change-Id: I8779f093696a16fbd455ad83c7ac8829e6c8fd41
Jira: CPOUI5FRAMEWORK-429