[FEATURE] Icon Explorer: Added appearance change option
New option is added to the Icon explorer application for theme changing,
alligned with the theme picker in Demo Kit.
Also, new utility is added called ThemePicker, which enables theme
changing and could be used in different applications.
Now, both the Demo Kit and the Icon Explorer are using the new utility
ThemePicker for their theme functionality.

JIRA: BGSOFUIPIRIN-5993
Change-Id: Id44f8d3a4b2fa65825bc53d781aa47d859ad98ab