[INTERNAL] SelectionController: introduce stable keys
Introduce 'stableKeys' similar to comp controller 'stableColumnKeys' feature
to avoid reordering of the first column.

The feature has been introduced in the SelectionController and will exclude
any given stable key from the p13n UI while calculating the delta as if
the key was present. The feature can also be used for other controls if
needed.

Jira: CPOUIFDENVER-1861
Change-Id: I7da49074a65a93139aed8793af23794cdbd92b6f