[FIX] sap.ui.model.odata.v2.ODataModel: Normalize entity key from location header for function imports
The location response header for the function import may contain encoded characters like
"%2A" for "*". These characters are decoded on importing an entity when the entity key
derived from __metadata.uri in the response is normalized. Normalization also has to be done
for the entity key in the location header so that messages have the correct full target.

PS1: Implement
PS2: Refer to correct BCP in commit message

Change-Id: I449407b30c99512acf8687a59fb326b718d31b34
BCP: 2270118627