[INTERNAL] demokit/walktrough: Adjust custom formatter to best practise
When it comes to asynchronous loading, it is not guaranteed anymore, that
one can access models via the view instance.

e.g. "this.getView().getModel() might return undefined because the view
might not be attached to the parent yet, and thus the view can't inherit
a model from that parent.

However, it is safe to go via the component to access models.

JIRA: CPOUI5FRAMEWORK-437
Change-Id: Ifaa75fab5da1af38db143155f2d5458db90034a7