[INTERNAL][FIX] PersistenceProvider: manual VM registration
When using the PersistenceProvider through the DPR, it will not have a
direct parent, hence the context needs to be provided manually.

In addition, the VariantModel needs to be reattached, and reregistered
by the VM control - introducing a ui5-restricted API on the VM to resolve
this issue, by re-initializing the VM created by the PersistenceProvider,
whenever the PersistenceProvider model context changes.

BCP: 2380009272
CR-Id: 002075125800000398582023
Change-Id: Ida31dda32519b3ecffba9764158c8520c6318dc0