[FIX] sap.m.URLHelper: Fixed redirect method
URLHelper redirect method contains code, which is dedicated to a problem
with window phone devices.It is causing a following issue on Safari
browser:
When you open a link with redirect, bNewWindow = true
(sap.m.UploadCollection) samples, it opens the link in a new tab and the
back button is disabled.
It is now fixed.

BCP: 6450822022
CR-Id: 002075125800001200862023
Change-Id: Ic2eb1ea843a05b1124239c2f1457684b44131748