[INTERNAL][FIX] TreeBindingProxy: Remove intrusive error when V4 flag not activated
If the OData V4 flag is not set to true and the TreeTable is bound against a OData V4 binding, an error was thrown.
This leads to several unwanted errors, when e.g. the table is bound against a valid V2 model after view initialization.
Throwing an error in this case is intrusive and interrupts other controls, leading to unwanted exceptions.
This error was changed to a logging error.

BCP: CS20230004690437
Change-Id: I407ff0b733527ed6ce2f9d0aa60989d157eab7bc