[INTERNAL] MDCTable: Change focus handling with filter info bar
The p13n dialog can be opened by pressing the filter info bar. When closing, the dialog (actually core.Popup) tries to restore the focus on the element that had the focus before opening the dialog. If all filters are removed though, the filter info bar is made invisible and cannot be focused anymore. The focus now moves to the document body.

The workaround that is applied in the table to move the focus to the table instead is no longer applicable with If1ce94134c1543cb145e5b0a5e674db0ca0a24ae and is replaced.

Change-Id: I98de5b3a1a45a707a56c8b7741b390f6c0d4b572