[FEATURE] sap.m.DynamicDateRange: Implement public static toDates method
With this method we provide a static public API as part of the DynamicDateRange
control that will replace the current toDates method in the DynamicDateUtil.
Currently the DynamicDateUtil.toDates returns an array of UniversalDates
which is private type for ui5. The new method in DDR will now return an
array of JS Dates or UI5Dates. The method was implemented as part of the
DDR as part of the isolation and deprication of the DynamicDateUtil.

Change-Id: I8c7232d36e64eacbbf0d698d34b1df04dc7770c5