[INTERNAL] SortFlex, GroupFlex: improve appliance for performed removes
In case a sorter/grouping for is for some reason not existing to the time
a similar "removeXXX" change is being performed for the affected key,
it needs to be ensured that the change is marked as not applicable
to avoid Array#splice being called with -1, which breaks the existing
housekeeping.

Possible scenarios in which this might occur is for example a change from
a deeper layer, hence we need to ensure graceful appliance.

BCP: 2270143640
Change-Id: Idc7697ec36ca93064aa2b8d8b1e6c09d7206844c