[INTERNAL][FIX] TreeBindingProxy: Fix initialisation error of TreeTable if V4 flag is not set
The previous change removed the intrusive errors thrown by the TreeTable to allow other controls to still be rendered. This resulted in a generic error being thrown, as initialisation still failed.
This was due to the case, that getBindingContexts() did not return an array in an error case.

Added return values for methods, were a return value is expected.

BCP: 002075129400000166292023
Change-Id: I361f293dee163f6844b1f6cd3fc46f8b77dc7fef