[INTERNAL] sap.ui.fl: fix condensing of create/move + destroy
When both Create and Destroy changes are available, both should be deleted. Due to a bug only the destroy got the condenserState 'delete'.

There is currently no control using the destroy classification of the Condenser,
that is why there is no test yet. This will be provided with the enablement of mdc controls.

Change-Id: Ic3bdc15965b7283d4514a0fe52d6cc95443eabbb