[INTERNAL] sap.m.DynamicDateRange: Use UI5Date instead of native Date
Returning DynamicDateRange and the controls that the control works with
to the state before integrating the timezone feature.
Also consuming the UI5Date object and replacing the usage of JS Date with
it. The new return types are also UI5Dates when used with the toDates method
which is public static method created in the DynamicDateRange.

JIRA: BGSOFUIBALKAN-6426

Change-Id: I98d7ce3f4a0ec2ec2c0cf6490bee18ac40de6c21