[INTERNAL] sap.ui.fl: enable custom aggregation for index-relevant condensing
previously index-relevant condensing was only enabled for UI5 aggregations.
Now controls can specify a location where an array is which will then be used as a custom aggregation.

Additionally items in aggregations don't need to have a getId function anymore,
a property name can be given in the condenser info that will be used instead.

Tests will be provided as soon as the mdc controls have adapted the new concept

JIRA: TEAM42-248
Change-Id: Ie95875fb6112aa17abf338ca48820310863037e2