[FIX] sap.m.TimePicker: Ignore TimePicker rules in adaptation mode
As the MaskInput is not relevant for UI Adaptation don't need a stable ID,
but for the support rule to know this they have to be marked.
Since they are all in an aggregation, it is ignored in the
designtime of the control.

Tested: Using the Support Assistant with rule for stable ID. Tests should
be performed in applications or snippets, as the demokit has a workaround
build in so that all IDs appear to be stable.

BCP: 2270191208

Change-Id: Iff918df3bb14ac3534826ecbb0289f29380055d6