[FIX] sap.m.DynamicDateRange: improve navigation in options list
ISSUE: As we have several options in Last X (days, weeks, months...), but
only LASTDAYS is part of the list, there was a mismatching with options
different than LASTDAYS( example: LASTWEEKS is part of the LASTDAYS page).

SOLUTION: Check if the options are one of the group and replace the search
in the list with either LASTDAYS or NEXTDAYS, so we can focus the correct
ListItem in the list.


BCP: 2380000923

Change-Id: I83196e88603e27345832663046689ac537ff57e6