[FIX] sap.m.OverflowToolbar: Issue in resizable Dialog fixed
Problem: When the OverflowToolbar is placed in a resizable Dialog, and
the Dialog is getting bigger, then closed, then opened again - some
controls of the OverflowToolbar go into the Popover, without the need to
(there is enough space for them).

Solution: The problem happens because the OverflowToolbar works with
cached sizes of its content controls. When some of them have relative
sizes (in %), if the Dialog is resized to a bigger size, those controls
also take more space. When the Dialog is closed, it just removes all its
content (the OverflowToolbar and its controls) directly from the DOM.
When the Dialog is opened again (with initial fixed size, comming from
contentWidth/Height property), the OverflowToolbar is making its
calculations, based on the previously cached sizes of the controls
inside it (bigger sizes, due to the bigger size of the resized Dialog).
At this point, the OverflowToolbar considers that it does not have
enough space to render all its controls in the toolbar, so it moves some
of them in the Popover.
The OverflowToolbar does not "know" when it is removed from the DOM.
That is why, now we just clear the cache in onBeforeRendering, if at the
moment of invocation, there is currently no DomRef (meaning: it is the
first rendering, so nevermind - there is still no cache anyway; or - the
DomRef was removed from the DOM for some reason, like it happens, when
the Dialog is closed).

BCP: 2380024979
Change-Id: I7204dfb51b8906396111472ce2bca08787da9b08