[INTERNAL][FIX] sap.ui.rta: Safeguard DesignTime access in elementModified callback
For performance reasons, some checks in the elementModified callback of
the plugin are debounced to avoid expensive recalculations.
These checks access the DesignTime which might have been destroyed
while waiting for the debounce timeout.
Since no further plugin or overlay updates are required after the DT was
destroyed, it is safe to ignore these queued calculations.

Change-Id: I81fa0914d73d3553c41e20e904c8d1faf6658d46
BCP: 2380026607