[FIX] sap.m.ObjectStatus: Correct alignment in Horizon
Previously we had padding left/right because of the missing
left/right border when in active state.
Now we're removing it because it brakes the alignment when in
forms and other layout controls.
This is in regards to the Focus Handling which is correctly
implemented and according to our visual designers this has to
be handled in the layout.

Change-Id: I27a85ee38f9a2a51dbd9ffb869706792828966e9
BCP: 2380026092