[INTERNAL][FIX]sap.ui.fl/rta: Errors discarding draft
When discarding a draft, an extra "toExternal" call is triggered even
when the hash technical parameters don't change, which causes a duplicate
ID issue when loading the component.

Also: the Fiori header can be destroyed before the RTA toolbar finishes
hiding - which causes an error later when removing the "hidden" style
class from the Fiori header.

For the toolbar, this is a temporary fix to ensure the release stability.
A better solution with a proper lifecycle flow will be prepared later.

Change-Id: Ie2ef8a00c030ec15349347dbb16ab8926a6ffa7c
BCP: 2380022992