[INTERNAL] sap.ui.fl: adapt revertAddedControls to only destroy if available
Previously, before a control was destroyed, it was removed from the aggregation.
This had the effect that the corresponding overlay was not destroyed because the
destroy was called on a control without parent and connection to the RootControl.
As the destroy automatically removed the control from the aggregation this step is skipped.

BCP: 2380029095
Change-Id: Ibfe9ee196aaafe82e510e5be65815f835f8e32ec