[INTERNAL][FIX]sap.ui.rta: Fiori toolbar error
There is an error when the Fiori toolbar is destroyed before the "hide"
process is finalized, because the "destroy" explicitly removed the
reference to the fiori header which needs to be made visible again when
the Fiori toolbar is hidden.

With this fix the reference to the fiori header is no longer explicitly
deleted.

Change-Id: I8332b9474f06b3b74b8eba29a9d215d0afec51b3