[INTERNAL] sap.m.IllustratedMessage: New sample with custom illus set
There is a requirement to demonstrate the sap.m.IllustratedMessage being
used with a custom illustration set (a one which is not a part of our
framework).

A new sample in the Demo Kit called "Illustrated Message in a Page with
Configurable Properties (Custom Illustration Set)" is introduced where
the requirement is fulfilled.

JIRA: BGSOFUIPIRIN-6046
Change-Id: Ic7f28d739b1f4c604c41e2a2c2d8d1d19bfdc689