[INTERNAL] sap.ui.core: UIComponent_unavoidablySync.qunit.js
The test expects the onInit hook of the nested view controller
to be executed. The Component.create call only awaits the loaded
promise of the root control, but does not wait for the nested view.
Since the nested view gets created async, we need to wait for the
loaded promise here as well.

Async flag on the nested view was introduced with
fcd99dc9f6e1abbc09cf99f3e340ea659efe3f84

Change-Id: I2527e23d6f303025950b286c83540c8c82347779