[FIX] sap.uxap.ObjectPageLayout: Misaligned actions are fixed
Problem: When ObjectPageLayout is in iconTabBar mode and there is only
one SubSection in a given Section, and it has some actions, the actions
overlap the content of the SubSection.

Solution: This happens because of a previous merging of the title of the
Section with the actions of the SubSection (new feature). However, in
iconTabBarMode when we do not show the title of the promoted SubSection,
the actions are overlapping the content, because the SubSection header
does not have height. Now we check for iconTabBar mode to disable the
CSS overrides for merging title with actions in this case.

BCP: 2380007216
Change-Id: I66cfe182b318a9a4638a2bf066000eb353e83d61