[FIX] Avoid assertion violation in RenderManager#openEnd
The rendering API openEnd() does not expect an argument. In some cases,
during the refactoring from apiVersion 1 to apiVersion 2, an argument
was provided (e.g. either the name of the opened tag or a single ">",
typically a left over from the old write() call).

These arguments cause an assertion violation and need to be removed.

Related to https://github.com/SAP/openui5/issues/3691.

Change-Id: I9b70525bbf9a2629151fb2a469e68847691bf522
BCP: 2380031904