[INTERNAL] Icon Explorer: New button for copying icon symbol
There is a requirement for a new button on each of the GridListItems.

Previously in order to copy the icon symbol you have to click on any
GridListItem and then click one more time in the right side panel in
order to do so.

We keep this feature, but we now present the option to copy the symbol
directly with a single click by clicking on the copy icon next/right to
the favorite (star) icon on each of the GridListItems.

JIRA: BGSOFUIPIRIN-4024
Change-Id: Ic67f58f24fa17368cb9a78102917c25ee018b3e5