[INTERNAL] odata.v4.TableDelegate: Clear selection when updating the binding
When changes are made in the binding with multiple API calls, the binding fires a change event with the consolidated reason "change". The information that there is a sort or filter change is lost, hence the inner GridTable does not clear the selection. The changes could affect the indices and make the current selection invalid. Therefore, the delegate has to clear the selection.

Change-Id: Ia30e080ec5cecc3a3ee7d4e14188d44fedc000f5
BCP: 2280186339
BCP: 2380013422