[INTERNAL] sap.m.SegmentedButton: legacy-free UI5 adjustments
The 'buttons' aggregation is deprecated and supposed to be removed in legacy-free UI5, but it is still used internally. Also, the default aggregation is 'buttons'.

As agreed, the legacy-free UI5 transformer will remove the defaultAggregation and change the 'buttons' aggregation from deprecated to hidden (in a separate change).

JIRA: BGSOFUIBALKAN-6518

Change-Id: Icd27286f6de4d723955b8d6b1775af7ca9a6d8db