[FEATURE] Integration Cards: New experimental property previewMode added
Available options are "MockData", "Abstract" and "Off".
The CardPreviewMode is meant to be used when the end user is choosing or configuring a card.
In preview mode "MockData", data is loaded according to the "data/mockData" manifest configuration.
If such is missing, the data is loaded from the original configuration, and informative message is displayed to the end user.

In Card Explorer:
- Added samples under explore/integrate/preview/
- Added sample features/data/mockData
- Added documentation in learn/data and integrate/preview

JIRA: BGSOFUIRODOPI-2957
Change-Id: If87ea3c7c5f97f8c06658074315fd13815e3aaed