[FIX] core/Rendering: Fix async step detection
Remove Interaction.notifyAsyncStep for sync renderPendingUiUpdates call.
Always call Interaction.notifyAsyncStep before async renderPendingUiPdates
call. For sync calls the callback in aFnDone will most likely not be
executed because there is no sRerenderTimer therefore the callback
for Interaction.notifyAsyncStep will not be executed and the interaction
counter is not in sync anymore.

Change-Id: If5e34bf7c1cceebcdf97cbebba94d054621eb3ad