[INTERNAL]: sap.m.Select: Provide editable input
Utils such as getFirstEditableInput are not able to find the sap.m.Select because the control is not using a native input elememnt as a focus target. However, the sap.m.Select has an internal input element, used for
accessibility mappings (e.g when submited by a form), which input can be
used for such purposes also.

With this change, we make the input visible in the DOM, yet not visible to the users and make sure to forward the browser focus to the focus target of the sap.m.Select.

BCP: 2380035090
Change-Id: Ic01b2519d1a387650deda9bf86f1de0b7999d909