[INTERNAL] Demo Kit: Removed redundant brackets from URL
Problem: Sometimes when we click on a link in the events tab of the API
ref which has brackets at the end ('[]'), we end up with non-existing URL.

Root Cause: Some times the event parameters types are marked as arrays
thus the brackets at the end. This results in an incorrect URL.

Solution: We force remove all brackets from the URL since brackets are
not recommended for URL structure. Once we remove them, the links work
properly.

BCP: 2380008679
Change-Id: I5edd20e3eee08a97c3511171d9136c2511089a42