[FIX] sap.m.ObjectAttribute: Tooltip correctly shown
Issue:
The tooltip of the active ObjectAttribute was only displayed
over the text and not the title.

Cause:
pointer-events: none is set to it thus no tooltips could be shown.

Solution:
The title has now have pointer-events: auto.
This was already made for the text for cases when ObjectAttribute
is inside ObjectListItem, it occupies 100% width and overlays part
of the parent control resulting in non clickable section there.

Change-Id: Id5a5cf309c7caa01e94f7a3f82d92b1260edee63
BCP: 2380042250