[FIX] sap.uxap.ObjectPageSubSection: Override parent aggregation name
Problem:
The subSections of a section are laid-out using a Grid control,
so they belong to that Grid's "content" aggregation, even
if externally they belong to the "subSections" aggregation
of their containing section.
We already overwrote the public "getParent" function to ensure that
only the public parent can be accessed from outside.
The problem is that some private API (e.g. sParentAggregationName)
still show the actual aggregation and cause issues when moving the
subsection from one parent to another with drag'n'drop during RTA

Solution:
Override the sParentAggregationName to match the publicly visible
parent.

BCP: 2380040241
Change-Id: Icc3caecd67ccd14c3d8089e2df01f5c0d9817fb8