[INTERNAL] sap.uxap.ObjectPage: Improve header label
It should be avoided to use the word “header” on its own. The reason is that without context, it can mean either something like “header area” or “header data” and the translations are different. So to make understanding easier, the recommendation is to use the relevant compound also in English.

CR-Id: 002075125900000048812023
BCP: 002075129500006802072022
Change-Id: I17527a6ee8ea4edcfe240f6a79c46cfbbe034ba6
(cherry picked from commit a8fcfebcc577a93bb98d762c044c6e4787f0f386)