[INTERNAL] sap.uxap.ObjectPage: Improve header label
It should be avoided to use the word “header” on its own. The reason is that without context, it can mean either something like “header area” or “header data” and the translations are different. So to make understanding easier, the recommendation is to use the relevant compound also in English.

Change-Id: I17527a6ee8ea4edcfe240f6a79c46cfbbe034ba6