[FIX] sap.m.DatePicker: Switch hours when minDate is after maxDate
Issue: When minDate comes after maxDate the dates are switched but the hours are incorrect, resulting in out-of-bounds date selections.

Solution: The hours for minDate and maxDate are now set to the beginning and end of the days respectively.

BCP: 2370001157
Change-Id: I10e67fa6af100f5579305e900e53510ae4f49e2c
(cherry picked from commit 73122f8582e6a207aed4c4f9762365ff482b5f36)
CR-Id: 002075125900000075192023