[INTERNAL] Demo Kit: Improved survey redirect text
Problem: When we are redirected to a survey from the upper-left
feedback button, and we end up in the new page where the survey should
be performed, we always show SAPUI5 for the name of the product. Even
when we are navigating from the OpenUI5 SDK.

Solution: We now display SAPUI5/OpenUI5 as the name of the product so
the newly navigated page title is always valid.

Additionally the end of the URL is now properly encoded.

JIRA: BGSOFUIPIRIN-6065

Change-Id: I1c932d826f1faddd1850a41187be77124336f7a9