[INTERNAL] Control: Invalidation handling during the rendering
 - The invalidation bubbling during the rendering might cause double
rendering on the parent if the child has not DOM output.
 - If a parent control is in a rendering phase then the invalidation of
its children should not necessarily bubble up the control hierarchy if
the rendering of the child control has not yet been started. In this
case the invalidation of the child control cannot simply be ignored
since the renderer code must not necessarily render the child control
after the invalidation. If the parent renderer code does not render the
child control after invalidating it then the UIArea still keeps track of
invalidated controls for the next rendering cycle.
 - With this change we will not lose invalidations but we can ignore the
ones that bubbles and invalidates the parent.

 - This change is important to be able convert apiVersion:2 controls to
apiVersion:4 easily. At the moment controls try to avoid invalidations
during the rendering with th bSuppressInvalidate flag of protected
setter calls. This avoids using public APIs. Besides that the
apiVersion:4 only works with valid rendering encapsulation, basically if
a control property is changed the control must be invalidated to be able
to signal the RenderManager to render it.

Change-Id: Ide0d10785c08df032b20dbe5683856a42adfbf9f