[INTERNAL] mdc/Table: copyProvider aggregation added
 - A dedicated copyProvider aggregation is now available for the
mdc/Table.
 - In contrast to mdc/Table paste settings (enablePaste/showPasteButton
properties and paste event) there is no copy settings that belongs to
the table. The copy related settings must be handled via CopyProvider
plugin defined in the dedicated aggregation.
 - CopyProvider is now capable to extract data if the plugin owner
implements getColumnClipboardSettings method and returns the desired
sap.m.plugins.CopyProvider.ColumnClipboardSettings type
 - CopyProvider also helps consumer to create a copy button and it
manages the enabled state of the button depending on the the enabled
state of the plugin itself.
 - The visible property of the CopyProvider manages the visibility of
the Copy button which can be created with the public #getCopyButton
method.
 - clipboardSettings option is added to mdc/PropertyInfo. This can be
used to disable the copy with setting it to null or define the template
for complex properties.
 - PropertyHelper#getColumnClipboardSettings method is implemented to
bring the clipboard settings of the column to extract the data for the
clipboard.
 - mdc/Table and standalone CopyProvider samples are changed with the
dedicated copyProvider aggregation.

Jira: CPOUIFTEAMB-2118

Change-Id: I000b7fc348f0359871c4a45cf12acc2286384bdc