[FIX] Integration Cards: Loading placeholder of Analytical card is now shown
In order to create loading placeholder, the content needs its manifest configuration

Problem:
Configuration is set and applied after the dependencies of the content are loaded, which is too late,
and especially noticeable in Analytical and Timeline contents

Solution:
Split the configuration setting and application in 2 steps:
1) set the configuration during content creation (immediately) in order to create a placeholder
2) apply it later, when the dependencies are loaded

Even then, if the loading provider of the content is JSON, the placeholder would not show again.
Now each card part decides when to show its placeholder, for the purpose of which the "isLoading" methods are refactored.

BCP: 2380027532
Snippix: 817509, 425767
Change-Id: I7b72720d14c5ccbf93937bab4a662e89aa1c6298