[INTERNAL] sap.ui.integration.cards.AdaptiveContent: Properly maintain ready state
Continuation of 6a395020f0b2ee96852297bf462f6d9b5370ca9f
While the card is loading its dependencies and data, it should be showing a loading placeholder.
Now the "ready" state is properly maintained, and thus a loading placeholder is correctly shown.

Also, the _ready event is now fired when the card is really ready. This makes timeouts in tests redundant.
Pending rendering can be forced with Core.applyChanges()

Change-Id: I6ec96975ce027d5ab753b36fe602b6f039740da1