[FIX] sap.m.DynamicDateRange: Last/Next options date format is appropriate
Issue:
When instantiating a DynamicDateRange without LastMinutes/LastHours,
but with other Last options, the formatiing would either show hours,
minutes and seconds when it's not needed, or not show it when it
should be there. The issue had to do with searching for options by
index even when some of these options are not present.

Solution:
Added checks that better take into account what options are actually
present in the control when determining whether a particular option
requires date or datetime formatting.

BCP: 2380039217
Change-Id: I736485d5b3e7f0f08f0332af274e836e51e82b43