[INTERNAL] Testsuite adjustment for sap.ui.integration lib
The test for 'TimelineCard' has a dependency to the 'suite-ui-commons'
library, which is only available in the distribution layer. Hence, this
test shall not be executed in the runtime scope, but only in the
distribution layer scope.
In turn, all other tests without any dependency shall only be executed
in the runtime scope and be excluded from the distribution layer scope.

Change-Id: I750de3ef70a9303003b914126270dc30e9a6f4a2