[FIX] ui.table: Navigation(Action) column must be in the scroll viewport
 - The min-width calculation of the Table converts rem width property
values of columns to pixel and this migh cause a rounding problems. At
the end for the large number of columns there might be less then a half
pixel difference. This change takes 1px width as an epsilon value to
determine whether the dummy column is visible or not.

BCP: 2370010378

Change-Id: I8adb9eb44109cd658ed1fac7129b5e59b40079fd