[FIX] sap.f.DynamicPageTitle: Overflow style rule adjusted
Problem: The overflow style rule which is applied to the
sap.f.DynamicPageTitle's actions aggregation causes visual issues
when the actions are rendered in a toolbar. If the actions are
of type sap.m.Button, the buttons are rendered normally, but when
they are being hovered, on Horizon theme, the right most box-shadow
is not entirely visible.

Motivation: There is a base overflow rule which
is applied to the DynamicPageTitle's actions aggregation.
It is defined in Toolbar.less. So in that case,
the overflow rule could be overridden.

BCP: 2370003608
Change-Id: I816814e5539686c9dc45e234bf19f0de84f45144