[INTERNAL] sap.ui.fl: fix condensing with multiple customAggregations
with standard aggregations, an ID is always unique per control.
But with custom aggregations the same ID could be used in different aggregations,
as not necessarily real controls are used. To enable this properly the
target aggregation was added to the internal condenser map.
This way the same ID can be used in different custom aggregations.

Also fixed a bug where the condenser state was not set to update for already persisted changes

BCP: 2370011022
Change-Id: Ia3c9667e042b780e8e82422f296de326f692c907