[FIX] sap.uxap.ObjectPageHeader: prevent duplication of buttons
Problem: When visibility of "action" buttons is updated, in some cases
the buttons are duplicated.

Solution: This happens for the buttons, which are initially in the
overflow. So when there visibility is updated to the same value (true),
they appear in the toolbar area as well.
Now we call _adaptLayout upon visibility update, so that after
calculations, all buttons would be shown only on one place - either in
the overflow or in the toolbar area.

BCP: 2380035583
Change-Id: Iaf5fd8bdf10dc5177f1090a9c6c64d135ae40c01