[INTERNAL] Introduction of "sap/base/config"
This change introduces the technical foundation for future
refactoring of the "sap/ui/core/Configuration".

With this change we introduce two features:
- a new "sap/base/config" module, which serves as the main API
to retrieve configuration parameters
- a set of default provider modules, which read the different
configuration sources

The new "sap/base/config" retrieves configuration parameter
values via the mentioned providers in the following priority:

1. URL parameters
2. meta tags defined in the page's HTML
3. the bootstrap's dataset
4. the global "sap-ui-config" object

Concerning "sap/base/config" get() API:

The "sap/base/config" module has a get() function which accepts the
parameter-name and the data type.
The parameter-name must be camel-case starting with a lower-case letter
followed by any letter or number.

- The data type is used for automatic type conversion, e.g.
boolean, string, ...
Additionally an enum can be given which will be used to check if
the defined parameter value corresponds to a valid enum value.

- The default value is returned in case no parameter value could be
found in any provider source. If no default value is provided, the
internal default value defined in "sap.base.config.Type" is being used.

- an "external" flag can be set,
which at the moment denotes if the URL parameters should be checked
In the future additional providers may be marked as "external" and
thus can either be included or excluded from parameter retrieval
via the get function.

Concerning compatibility:

To stay compatible with extisting minimal bootstraps which manually
include use the "ui5loader" and the "ui5loader-autoconfig",
we introduce our new config and provider modules as raw content into
the ui5loader-autoconfig.
This way we ensure that the new configuration modules are
available as early as possible and can directly be used inside
the ui5loader-autoconfig itself.

Change-Id: Id3ac21255db4481fff5d230a6ac017a6684a536f
JIRA: CPOUI5FRAMEWORK-461