[INTERNAL][FIX]sap.ui.fl: Duplicate Variant Model for app Component
If multiple embedded components are created in parallel, due to the
asynchronous nature of the Variant Model initialization, the model was
being created and initialized multiple times. This caused an error when
registering and deregistering navigation filters.

Now the embedded component always waits for the outer component and
reuses the VariantModel from there, but never creates an own VariantModel

BCP: 2370020409
Change-Id: Ia4f12f87390ed21824d670edc33e47ef81da5a78