[FIX] Integration cards: SkeletonCard can now be refreshed and ComboBox value will always be synchronized
The SkeletonCard does not get rendered, so the means we have for
triggering data requests based on whether it's being shown in the
viewport did not trigger. This did not allow the control to be refreshed.
With this fix, the SkeletonCard can now be refreshed.

Additionally, for the same reason, the Object card will now keep the
internal ComboBox selectedKey and value properties in sync.

JIRA: BGSOFUIRODOPI-2966
Change-Id: Ie6c52a2bf2ad9c6e0f81d8e6d89ea362e3d016f8