[INTERNAL] sap.m.SegmentedButton: prepare removal of "selectedButton"
The implementation of the SegmentedButton internally depends on the
deprecated association "selectedButton". A future release of UI5 might
convert this association to a hidden one. To prepare the implementation
for such a step, the necessary getter is added as code instead of
relying on the generated getter for the (still) public association.

Change-Id: I78757eb4485138f4d407ea351c18f6f8adb4150e