[INTERNAL] ListFieldHelpItem: move to FixedListItem
sap.ui.mdc.field.ListFieldHelpItem is placed in the "field" node, but belongs to the FixedList, so it should be placed in the "valuehelp" node. Also the naming needs to be changed.

So a new FixedListItem is creaded what is used in FixedList. The old ListFieldHelpItem inherits now from the new FixedListItem. So it is sill working. It is now deprecated and will later on deleted.

Change-Id: Icdf02a375841e1cdce3f81384c0f9110258bca0f
JIRA: CPOUIFDENVER-1892