[FIX] sap.f.DynamicPage: Fix growing Table
Problem: When there is a Table with growingScrollToLoad, placed in
DynamicPage, at some point the Table stops growing and does not show all
its items. The problem comes from a CSS style - overflow-y: auto, set on
the content wrapper of the DynamicPage, which now becomes scrollable. As
a result, the actual scrollable wrapper of the DynamicPage
(ScrollDelegate) is not scrolled and the callback of the Table, which
shows more and more items upon scroll, is not called.

Solution: Removing the overflow-y: auto style and instead of it, remove
the fitContent class of the DynamicPage, when scroll is needed (no
matter of the user setting of the fitContent property). This way, we fix
the bug, which required the overflow-y: auto to be set on the content
wrapper of the DynamicPage: AnalyticalTable, placed in the DynamicPage,
which has a minimum height (related to the minimm number of rows to be
displayed in the table) and which in some cases gets overlapped by the
footer of the DynamicPage (when the height of the page is not enough to
show all its content without scroll).

BCP: 2380041732
Change-Id: Idafd341cd48a210b860d845b49d55ec76369fe98