[FIX] sap.m.InputBase: Focus of busy input corrected
When an input is busy, the focus outline is not the one defined by the input,
but it is the outline of the local busy indicator.
Because of this, we should overwrite and adapt the local busy indicator focus styling
in order to fix the input fields case in the different themes.

Change-Id: I5c27d78fd0d259209f7bb2d243bf9d192e786c31
BCP: 2370014514