[FIX] sap.uxap.BlockBase: Fix setParent function
Problem: When view is parsed and the flexibility logic is trying to
retrieve and apply all changes, made on the ObjectPage, there is a moment
(before the first rendering), when getParent method of the BlockBase
returns null (when it is placed in an ObjectPageLazyLoader). As a
result, flexibility cannot find the View, containing the block, and
their logic does not work.

Solution: The problem happens between the view creation and the first
rendering, because just after the first rendering the BlockBase will
receive the internal Grid (created by the ObjectPageSubSection) as its
actual parent. In the meantime the parent is the ObjectPageLazyLoader,
but it gets destroyed at some point and setParent function is called
with the ObjectPageSubSection. However, due to the logic in the setter
(implemented because of another bug), if the setter is called with the
ObjectPageSubSection as a parent, the actual parent is not set at this
point (but later - when the internal Grid is created). Now we set the
parent to be the ObjectPageSubSection and remove the previous error that
had appeared in the past, by correctly calling setParent with the
aggregation name and also fix the overriden removeAggregation function
to working with vObject (instead of oObject), as in some cases it may be
called with iIndex, not with an object, which should be removed.

BCP: 2380044009
Change-Id: Ief3e60ce916f6c21248048badefc172062900346